+++ planacad/planificaciones/templates/index.html
@@ -13,6 +13,7 @@
   <body>
     <div class="container">
       <h1>PlanAcad</h1>
+      <p>test</p>
     </div>
   </body>
 </html>
